refactor: shell.py becomes flip.py — the prompt goes, the flip loop stays
Exiting the harness with no flip pending prints the resume hint and
returns to the OS shell; the tandem prompt loop is gone. Ctrl-]'s flip
machinery (ladder, held-back reports) is unchanged.
